+++ src/modules/glm/samplers/OrderedProbit.h
@@ -27,7 +27,7 @@ namespace glm {
     {
 	double const &_y;
 	double const *_cuts;
-	const unsigned int _ncut;
+	const unsigned long _ncut;
 	double _z;
       public:
 	OrderedProbit(StochasticNode const *snode, unsigned int chain);
