+++ src/modules/glm/samplers/REMethod.h
@@ -92,7 +92,7 @@ namespace glm {
 	 *
 	 */
 	void calCoefSigma(double *A, double *b, double const *sigma0,
-			  unsigned int m) const;
+			  unsigned long m) const;
 	/**
 	 * Calculates the likelihood for sigma, the vector of standard
 	 * deviation parameters for the random effects
@@ -106,7 +106,7 @@ namespace glm {
 	 * @see calCoefSigma
 	 */
 	double logLikelihoodSigma(double const *sigma, double const *sigma0,
-				  unsigned int m) const;
+				  unsigned long m) const;
     };
 
 }}
